Added the `Labels` property to the `TokenDBValue` struct
This patch extends the `TokenDBValue` struct in order to allow storing "labels" associated
with a user directly in the data structure, ie. without a need for secondary storage.

The primary motivation is fetching the user's teams from the Github API and storing
them with the user's token, so the teams can be checked in the ACL matchers.